Updates settings page
We move all of the individual forms to their own files for easier maintenance (as they are page components, they'll live here rather than in the ui/components file)

Each form allows us to edit the data on the client, and push an update to the server to store the updated data.

This should allow us to see the change reflected instantly on the client.

This separation should also make maintenance easier as each form has its own actions